added the things

+++ mysql/tran.sql
@@ -3,9 +3,10 @@ USE crud_app;
 
 CREATE TABLE entries (
   id INT AUTO_INCREMENT PRIMARY KEY,
-  amount DECIMAL(10, 2) NOT NULL,
+  amount INT NOT NULL,
   description VARCHAR(255) NOT NULL
 );
+
 
 CREATE USER IF NOT EXISTS 'crud'@'%' IDENTIFIED BY 'CrudApp@1';
 GRANT ALL ON crud_app.* TO 'crud'@'%';
